% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getFunctions}
\alias{getFunctions}
\alias{getFunctions,StQT-method}
\title{Returns slot \code{Functions} from an object of class \linkS4class{StQT}}
\usage{
getFunctions(object)

\S4method{getFunctions}{StQT}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{StQT}}}
}
\value{
A \code{\link{list}} corresponding to the slot \code{Functions} of the input parameter.
}
\description{
\code{getFunctions} extracts slot \code{Functions} from an \code{\linkS4class{StQT}} object.
}
\examples{
rules <- data.frame(domain = c('!is.na(Variable)','is.na(Variable)',''),
                    output = c('MediaVariable','Variable','MediaVariable'),
                    fun =    c('mean','identity','FunDelVar'),
                    input =  c('Variable','MediaVariable',''),
                    by =     c('Estrato','',''),
                    key =    c('Estrato','',''),
                    stringsAsFactors = FALSE)
T1 <- NewStQT(rules)
T1
getFunctions(T1)

}
