% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRules.R
\docType{methods}
\name{getRules}
\alias{getRules}
\alias{getRules,StQT-method}
\title{Returns slot \code{Rules} from an object of class \linkS4class{StQT}}
\usage{
getRules(object)

\S4method{getRules}{StQT}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{StQT}}}
}
\value{
A \code{\link{data.frame}} corresponding to the slot \code{Rules} of the input parameter.
}
\description{
\code{getRules} extracts slot \code{Rules} from an \code{\linkS4class{StQT}} object.
}
\examples{
T1 <- NewStQT(data.frame(
                domain = c("Next==Unit","","Value < 0"),
                output = c("Value_next","Value_next",""),
                   fun = c("FunAutoLink","FunDelVar","FunDelRow"),
                 input = c("Value","",""),
                         stringsAsFactors = FALSE))
T1
getRules(T1)


}

