% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.StQT.R
\docType{methods}
\name{length,StQT-method}
\alias{length,StQT-method}
\title{Method \code{length} for the class \linkS4class{StQT}}
\usage{
\S4method{length}{StQT}(x)
}
\arguments{
\item{x}{Object of class \code{StQT}.}
}
\value{
Number of rules of a StQT object.
}
\description{
\code{length} gets the number of rules of a StQT object.

This method overloads the method \code{\link{length}}.
}
\examples{
x <- 1

}

