% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewStQT.R
\docType{methods}
\name{NewStQT}
\alias{NewStQT}
\alias{NewStQT,character-method}
\alias{NewStQT,data.frame-method}
\alias{NewStQT,missing-method}
\title{Constructor de objetos de clase StQT}
\usage{
NewStQT(x)

\S4method{NewStQT}{missing}()

\S4method{NewStQT}{data.frame}(x)

\S4method{NewStQT}{character}(x)
}
\value{
Objeto de clase \linkS4class{StQT}.
}
\description{
Este constructor devuelve un objeto de clase
\code{\linkS4class{StQT}}
}
\examples{
x <- 1

}

