% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{methods}
\name{output}
\alias{output}
\alias{output,StQT-method}
\title{Output variables from a \linkS4class{StQT} object.}
\usage{
output(x)

\S4method{output}{StQT}(x)
}
\arguments{
\item{x}{Object of class \code{StQT}.}
}
\value{
Character vector with the variable names
}
\description{
\code{output} gets the output variables from a StQT object.
}
\examples{
x <- 1
}

