% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.StQT.R
\docType{methods}
\name{+,StQT,StQT-method}
\alias{+,StQT,StQT-method}
\title{Method \code{+} for the class \linkS4class{StQT}}
\usage{
\S4method{+}{StQT,StQT}(e1, e2)
}
\arguments{
\item{e1}{Object of class \code{StQT}.}

\item{e2}{Object of class \code{StQT}.}
}
\value{
Object of class \code{StQT} that binds e2 rules after e1 rules.
}
\description{
\code{+} joins two objects of class \linkS4class{StQT} in a
single object of the same class.

This method overloads the operator \code{\link{+}} and builds a new
\linkS4class{StQT} object joining both input objects.
}
\details{
Operator \code{+} is non-commutative for \linkS4class{StQT} objects. It is defined to
fulfill the equality: Tapply(Data,e1+e2) == Tapply(Tapply(Data,e1),e2)
}
\examples{
x <- 1

}
