% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.StQT.R
\docType{methods}
\name{show,StQT-method}
\alias{show,StQT-method}
\title{Show an object of class \linkS4class{StQT}}
\usage{
\S4method{show}{StQT}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQT}.}
}
\value{
Invisible \code{\link{NULL}}.
}
\description{
\code{show} displays the slot \code{Rules} of the input \linkS4class{StQT}
object. For too wide columns the values are cut and ... is displayed.
}
\details{
This method displays only the content of slot \code{Rules} from the input
\linkS4class{StQT} object. It is indeed the method \link[methods]{show} adapted
to class \linkS4class{StQT}.
}
\examples{
# A trivial example
show(new(Class = 'StQT'))

}

