% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQT-class.R
\docType{class}
\name{StQT-class}
\alias{StQT-class}
\title{S4 class of dataset transformations.}
\description{
Definition of an S4 class named \code{StQT} for dataset transformation.
}
\details{
The structure of the class \code{StQT} comprises 2 attributes:
\itemize{
\item The attribute \code{Rules}, which is a \link{data.frame} with the transformation rules.
\item The attribute \code{Functions}, which is a list of the functions used in the rules.
}
The rules are represented with the following fields in the \code{Rules} data.frame.
\itemize{
\item{domain}{Specifies where the rule is applied}
\item{output}{variable(s) de salida}
\item{fun}{función a aplicar}
\item{input}{variable(s) de entrada}
\item{by}{agrupación}
\item{order}{variable(s) por las que ordenar los datos}
\item{key}{key of the output variables}
}
}
\examples{
x <- new(Class = 'StQT', Rules=data.frame(domain=c("y<3",""),output=c("y=4","exponencial"),fun=c("sum","exp"),input=c("v","v"),by=c("x",""),stringsAsFactors=FALSE))



}

