% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.StQT.R
\docType{methods}
\name{[,StQT-method}
\alias{[,StQT-method}
\title{Extrae partes de un objeto de clase StQT.}
\usage{
\S4method{[}{StQT}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{objeto de clase \linkS4class{StQT} del que se van a extraer los
elementos.}

\item{i}{índices correspondientes a los elementos a extraer.}

\item{drop}{Incluido por coherencia.}
}
\value{
Objeto de clase \linkS4class{StQT}, que consiste en un subconjunto del
 objeto \code{StQT} de entrada.
}
\description{
\code{[} extrae o reemplaza partes de un objeto de clase \linkS4class{StQT}.
}
\details{
Se trata del método \code{[} para la clase \linkS4class{StQT}. Este método
obtiene subconjuntos del objeto de clase \code{StQT} especificado como entrada.
}
\examples{


}

