% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getFunctions}
\alias{getFunctions}
\alias{getFunctions,StQT-method}
\title{Devuelve el slot Functions de un objeto de clase StQT.}
\usage{
getFunctions(object)

\S4method{getFunctions}{StQT}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{StQT}} del que se desea
extraer las funciones.}
}
\value{
Una \code{\link{list}} con las funciones de transformación.
}
\description{
\code{getFunctions} devuelve el slot \code{Functions} del input
\code{\linkS4class{StQT}}.
}
\examples{


}

