% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRules.R
\docType{methods}
\name{getRules}
\alias{getRules}
\alias{getRules,StQT-method}
\title{Devuelve el slot Rules de un objeto de clase StQT.}
\usage{
getRules(object)

\S4method{getRules}{StQT}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{StQT}} del que se desea
extraer las reglas.}
}
\value{
Un \code{\link{data.frame}} con las reglas de la transformación.
}
\description{
\code{getRules} devuelve el slot \code{Rules} del input
\code{\linkS4class{StQT}}.
}
\examples{


}

