% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{methods}
\name{input}
\alias{input}
\alias{input,StQT-method}
\title{Input variables from a \linkS4class{StQT} object.}
\usage{
input(x)

\S4method{input}{StQT}(x)
}
\arguments{
\item{x}{Object of class \code{StQT}.}
}
\value{
Character vector with the variable names
}
\description{
\code{input} gets the input variables needed for a StQT object to work.
}
\examples{
x <- 1

}

