% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewStQT.R
\docType{methods}
\name{NewStQT}
\alias{NewStQT}
\alias{NewStQT,character-method}
\alias{NewStQT,data.frame-method}
\alias{NewStQT,missing-method}
\title{StQT class constructor}
\usage{
NewStQT(x)

\S4method{NewStQT}{missing}()

\S4method{NewStQT}{data.frame}(x)

\S4method{NewStQT}{character}(x)
}
\value{
\linkS4class{StQT} object.
}
\description{
This constructor returns an \code{\linkS4class{StQT}} object
}
\details{
You don't need to specify all rules fields, the constructor will
generate empty fields when necessary.
}
\examples{
NewStQT()    # Empty transformation (identity)
NewStQT(data.frame(
           output="VAR",
           fun="*",
           input="SD,SD",
           stringsAsFactors = FALSE))
# If we want to read the rules from a semicolon separated csv file:
# NewStQT("<filename>")

}

