% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQT-class.R
\docType{class}
\name{StQT-class}
\alias{StQT-class}
\title{S4 class of dataset transformations.}
\description{
Definition of an S4 class named \code{StQT} for data transformation.
}
\details{
The structure of the class \code{StQT} comprises 2 attributes:
\itemize{
\item The attribute \code{Rules}, which is a \link{data.frame} with the transformation rules.
\item The attribute \code{Functions}, which is a list of the functions used in the rules.
}
The rules are represented with the following fields in the \code{Rules} data.frame.
\itemize{
\item{domain}{Specifies where the rule is applied}
\item{output}{Output variables}
\item{fun}{Function to apply}
\item{input}{Input variables}
\item{by}{Variables to group by}
\item{order}{Variables to order by}
\item{key}{Key of the output variables}
}
}
\examples{
T1 <- new(Class = 'StQT', Rules = data.frame(
                             domain="",
                             output="VAR",
                             fun="*",
                             input="SD,SD",
                             by="",
                             order="",
                             key="",
                             stringsAsFactors = FALSE),
                             Functions = list(`*`=`*`))
T1
getRules(T1)
getFunctions(T1)

}

