% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQT-class.R
\docType{class}
\name{StQT-class}
\alias{StQT-class}
\title{Clase S4 de transformaciones de un conjunto de datos}
\description{
Definición de una clase S4 \code{StQT} de transformaciones de un conjunto de datos
}
\details{
La estructura de la clase S4 \code{StQT} se compone del slot \code{Rules}
que es un data.frame que contiene las reglas de transformación según el modelo
\describe{
\item{domain}{dónde se aplica la regla}
\item{output}{variable(s) de salida}
\item{fun}{función a aplicar}
\item{input}{variable(s) de entrada}
\item{by}{agrupación}
\item{order}{variable(s) por las que ordenar los datos}
\item{key}{key of the output variables}
} y del slot \code{Functions} que es una lista con todas las funciones utilizadas en dichas reglas.
}
\examples{
x <- new(Class = 'StQT', Rules=data.frame(domain=c("y<3",""),output=c("y=4","exponencial"),fun=c("sum","exp"),input=c("v","v"),by=c("x",""),stringsAsFactors=FALSE))



}

