% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.StQT.R
\docType{methods}
\name{[,StQT-method}
\alias{[,StQT-method}
\title{Extracts parts of an StQT object.}
\usage{
\S4method{[}{StQT}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{object of class \linkS4class{StQT} from which to extract element(s).}

\item{i}{indices specifying elements to extract.}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{StQT}, with the subset of rules.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{StQT}.
}
\details{
It is indeed the method \code{[} for the class \linkS4class{StQT}. This method
returns an object of class \linkS4class{StQT} whose rules are the specified subset
of the original rules.
}
\examples{
rules <- data.frame(domain = c('!is.na(Variable)','is.na(Variable)',''),
                    output = c('MediaVariable','Variable','MediaVariable'),
                    fun =    c('mean','identity','FunDelVar'),
                    input =  c('Variable','MediaVariable',''),
                    by =     c('Estrato','',''),
                    key =    c('Estrato','',''),
                    stringsAsFactors = FALSE)
imputa_media <- NewStQT(rules)
imputa_media
imputa_media[1:2]
imputa_media[3:1]

}
