% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TApply.R
\docType{methods}
\name{TApply}
\alias{TApply}
\alias{TApply,StQ,StQT-method}
\alias{TApply,data.table,StQT-method}
\title{Applies a transformation object on a StQ or data.table.}
\usage{
TApply(x, Tr)

\S4method{TApply}{data.table,StQT}(x, Tr)

\S4method{TApply}{StQ,StQT}(x, Tr)
}
\arguments{
\item{x}{Objeto de clase \linkS4class{StQ} o \linkS4class{data.table} con los
datos a transformar}

\item{Tr}{Objeto de clase \linkS4class{StQT} con las reglas de transformación
a aplicar.}
}
\value{
\linkS4class{StQ} o \linkS4class{data.table} con los datos transformados
}
\description{
\code{TApply} returns a transformed object according to the transformation rules.
}
\details{
There are three kinds of transformation rules:
internal functions, row insertion and column insertion.
\describe{
\item{Internal functions}{FunDelRow, FunDelCol and FunAutoLink so far. The first one deletes
rows in domain. The second one deletes columns in output. The third one copies the input variables from the table
linking according to the expression in domain.}
\item{Row insertion}{Row are inserted. This kind of rule is applied when there exists an assignment in the output.
Unassigned variables are calculated according output = fun(input). Any variable included in by is copied in the new rows.}
\item{Column insertion}{Se insertan columnas calculadas según el esquema output = fun (input)}
}
El parámetro domain restringe los registros en los que se realiza el cálculo.
El parámetro by aplica la transformación por subtablas según los valores de las variables
contenidas en dicho parámetro.
El parámetro order aplica una ordenación de la tabla (útil para incrementar la velocidad),
que se realiza previamente a la transformación correspondiente.
}
\examples{
x
}

