% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TApply.R
\docType{methods}
\name{TApply}
\alias{TApply}
\alias{TApply,StQ,StQT-method}
\alias{TApply,data.table,StQT-method}
\title{Applies a transformation object on a StQ or data.table.}
\usage{
TApply(x, Tr)

\S4method{TApply}{data.table,StQT}(x, Tr)

\S4method{TApply}{StQ,StQT}(x, Tr)
}
\arguments{
\item{x}{Object of class \linkS4class{StQ} or \linkS4class{data.table} with data
to transform}

\item{Tr}{Object of class \linkS4class{StQT} with the transformation we are going
to apply.}
}
\value{
\linkS4class{StQ} or \linkS4class{data.table} respectively with transformed
data
}
\description{
\code{TApply} returns a transformed object according to the transformation rules.
}
\details{
There are three kinds of transformation rules:
internal functions, row insertion and column insertion.
\describe{
\item{Internal functions}{FunDelRow, FunDelCol and FunAutoLink so far. The first one deletes
rows in domain. The second one deletes columns in output. The third one copies the input variables from the table
linking according to the expression in domain.}
\item{Row insertion}{Row are inserted. This kind of rule is applied when there exists an assignment in the output.
Unassigned variables are calculated according output = fun(input). Any variable included in by is copied into the new rows.}
\item{Column insertion}{Rows are calculated according output = fun (input)}
}
Parameter domain specifies the subset of the dataset where we are going to do the calculation.
El parámetro by aplica la transformación por subtablas según los valores de las variables
contenidas en dicho parámetro.
El parámetro order aplica una ordenación de la tabla (útil para incrementar la velocidad),
que se realiza previamente a la transformación correspondiente.
}
\examples{
x
}

