% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TApply.R
\docType{methods}
\name{TApply}
\alias{TApply}
\alias{TApply,StQ,StQT-method}
\alias{TApply,data.table,StQT-method}
\title{Applies a transformation object on a StQ or data.table.}
\usage{
TApply(x, Tr)

\S4method{TApply}{data.table,StQT}(x, Tr)

\S4method{TApply}{StQ,StQT}(x, Tr)
}
\arguments{
\item{x}{Objeto de clase \linkS4class{StQ} o \linkS4class{data.table} con los
datos a transformar}

\item{Tr}{Objeto de clase \linkS4class{StQT} con las reglas de transformación
a aplicar.}
}
\value{
\linkS4class{StQ} o \linkS4class{data.table} con los datos transformados
}
\description{
\code{TApply} returns a transformed object according to the transformation rules.
}
\details{
There are three kinds of transformation rules:
internal functions, row insertion and column insertion.
\describe{
\item{Internal functions}{FunDelRow, FunDelCol and FunAutoLink so far. The first one deletes
rows in domain. La segunda borra las columnas
especificadas en output. La tercera extrae datos de la propia tabla, enlazando según las
igualdades de domain, copiando el input en el output}
\item{Inserción de filas}{Se insertan filas. Los output que contienen una igualdad se asignan directamente
y el resto se calculan según el esquema output=fun(input)}
\item{Inserción de columnas}{Se insertan columnas calculadas según el esquema output = fun (input)}
}
El parámetro domain restringe los registros en los que se realiza el cálculo.
El parámetro by aplica la transformación por subtablas según los valores de las variables
contenidas en dicho parámetro.
El parámetro key aplica una ordenación de la tabla (útil para incrementar la velocidad),
que se realiza previamente a la transformación correspondiente.
}
\examples{
x
}

