% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TApply.R
\docType{methods}
\name{TApply}
\alias{TApply}
\alias{TApply,StQ,StQT-method}
\alias{TApply,data.table,StQT-method}
\title{Aplica una transformación a una tabla de datos.}
\usage{
TApply(x, Tr)

\S4method{TApply}{data.table,StQT}(x, Tr)

\S4method{TApply}{StQ,StQT}(x, Tr)
}
\arguments{
\item{x}{Objeto de clase \linkS4class{StQ} o \linkS4class{data.table} con los
datos a transformar}

\item{Tr}{Objeto de clase \linkS4class{StQT} con las reglas de transformación
a aplicar.}
}
\value{
\linkS4class{StQ} o \linkS4class{data.table} con los datos transformados
}
\description{
\code{TApply} devuelve una tabla de datos transformada.
}
\details{
Existen tres tipos de reglas de transformación básicas:
funciones internas, inserción de filas e inserción de columnas.
\describe{
\item{Funciones internas}{FunDelRow, FunDelCol y FunAutoLink. La primera borra
las filas que cumplen la condición especificada en domain. La segunda borra las columnas
especificadas en output. La tercera extrae datos de la propia tabla, enlazando según las
igualdades de domain, copiando el input en el output}
\item{Inserción de filas}{Se insertan filas. Los output que contienen una igualdad se asignan directamente
y el resto se calculan según el esquema output=fun(input)}
\item{Inserción de columnas}{Se insertan columnas calculadas según el esquema output = fun (input)}
}
El parámetro domain restringe los registros en los que se realiza el cálculo.
El parámetro by aplica la transformación por subtablas según los valores de las variables
contenidas en dicho parámetro.
El parámetro key aplica una ordenación de la tabla (útil para incrementar la velocidad),
que se realiza previamente a la transformación correspondiente.
}
\examples{
x
}

